requirements.txt
Django==1.7.4
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.2.1
psycopg2==2.6
static3==0.5.1
djangorestframework==3.0.5
